23 log energy value during traing (#25)
* Calculating average training energy
* Added continuous sampling until energy is lower then training energy
* Added generation loop

---------

+++ utils/Sampler.py
@@ -21,7 +21,7 @@ class Sampler:
         self.max_len = max_len
         self.buffer = [(torch.rand((1,) + img_shape) * 2 - 1) for _ in range(self.sample_size)]
 
-    def sample_new_tensor(self, steps: int = 60, step_size: int = 10):
+    def sample_new_tensor(self, steps: int = 60, step_size: float = 10.0):
         """
         Function for getting a new batch of sampled tensors via MCMC.
         Inputs:
